# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cafe Backend API",
        "description": "Dokumentasi API Cafe Backend",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Admin Transactions"
                ],
                "summary": "Ambil semua transaksi",
                "description": "Mengambil seluruh data transaksi beserta relasi order",
                "operationId": "6ffdc60927f92c379d6cbb20b9490aa8",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar transaksi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Transactions"
                ],
                "summary": "Konfirmasi pembayaran / buat transaksi",
                "description": "Membuat transaksi baru atau mengupdate transaksi lama menjadi paid, sekaligus mengubah order.payment_status menjadi paid",
                "operationId": "a1ad9683ef73b722dca65dc85d264edc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "payment_method"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "example": "cash"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pembayaran berhasil dikonfirmasi"
                    },
                    "409": {
                        "description": "Order sudah dibayar"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Admin Transactions"
                ],
                "summary": "Ambil detail transaksi",
                "description": "Mengambil satu transaksi berdasarkan ID beserta order terkait",
                "operationId": "b401477f64714b13d759a0ab26c10ffc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaksi ditemukan"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/transactions/{id}/refund": {
            "post": {
                "tags": [
                    "Admin Transactions"
                ],
                "summary": "Refund transaksi",
                "description": "Mengubah status transaksi menjadi refunded dan membatalkan order terkait",
                "operationId": "46be35f7cb76af7034eb2bd2c2138dca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Refund berhasil"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    },
                    "400": {
                        "description": "Transaksi tidak valid untuk refund"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT dengan format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}